Adding time to active contracts
Once a contract is established between a player and a Kingdom island it must be maintained for a period of time (currently set to a full turn of the moon, 8 phases). The contract requires delivery of one good per turn which is assumed taken by the Kingdom ships along the trade route. This update:
* Adds a countdown of the remaining time required for an active contract to be completed.
* Updates the contract dashboard (on right sidebar) to show the remaining time for each contract.
* Reduces a player's goods being each turn based on the contracts that are active (essentially using up the goods being created by the resource each turn).
* Adds the resource tile and the contract trade route to the contract array for later use.
This update:
* Adds