Add files via upload
Now that a team can only have one contract with an island, it is useful to add a team colour marker next to the island once a contract is complete. This makes a quick reference on the board to illustrate that the team already has completed a contract with that island and will remind them not to initiate another contract with the same island. 

board.js
* Added the method which adds a mark in the appropriate harbour to show team has completed contract with that island (the harbour marks align with the position of the team settlements).

contracts.js