Automated ship movement - Avoiding pirat ships
Prior to this commit computer opponents could search for Resource tiles but they did so oblivious to any danger from pirate ships. This commit adds a cost for a move tile being close to pirate ships. Initially the cost is just based on the distance to each pirate ship. This cost reduces the ranking points when determining the best destination to move to.

movement.js
* paintFindPath - Method updated with distance of pirate ships from tile 
* postPaintFindPath - New method. Finds last active tile of path to destination.

automove.js
* rankDestinations - Method updated with points cost of move being close to pirate ships (uses last tile - i.e. tile actually moved to - rather than final destination which may be a couple of moves away so pirate presence there is not relevant).