Computer opponent - improving resource decisions
The previous ranking method for determining which island to move towards and search was quite simple. This improved method combines the probabilities of finding each resource with a points rating for each resource to give an overall resource ranking for each island search.

This was necessary to improve decision making by computer opponents, particularly relating to choosing between undiscovered islands and unclaimed resources towards the end of the resource search.