Automated scrolling
Graphical improvement - game now scrolls to ship that is about to move. There is a delay before the scrolling, before tiles are activated, and before the ship starts to move so the game does not feel rushed.
Scrolling limited to upper and lower edges of board. 
Restriction on minimum scrolling to prevent jumpiness.
Implemented using async/await - aim is to slowly refactor to async rather than callbacks.

Changes mainly in: movement.js: shipTransition